remove pre-commit-validate-config and pre-commit-validate-manifest
