+++ pre_commit/commands/validate_config.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
+from typing import Sequence
+
 from pre_commit import clientlib
 
 
-def validate_config(filenames: list[str]) -> int:
+def validate_config(filenames: Sequence[str]) -> int:
     ret = 0
 
     for filename in filenames:
