+++ setup.cfg
@@ -34,8 +34,6 @@ exclude =
 [options.entry_points]
 console_scripts =
     pre-commit = pre_commit.main:main
-    pre-commit-validate-config = pre_commit.clientlib:validate_config_main
-    pre-commit-validate-manifest = pre_commit.clientlib:validate_manifest_main
 
 [options.package_data]
 pre_commit.resources =
